+++ __tests__/wizard.open-via-button.test.js
@@ -14,7 +14,7 @@ beforeAll(async () => {
   global.localStorage = dom.window.localStorage;
   global.Chart = function(){};
   window.HTMLCanvasElement.prototype.getContext = () => ({});
-  window.api = { libs:{}, bus: mitt(), version: () => '0' };
+  window.api = { bus: mitt(), version: () => '0' };
   renderer = await import('../src/renderer/renderer.js');
 });